+++ web/api/admin/admin_car.py
@@ -12,17 +12,31 @@ from api.models import db, Car, PersonType, CarSchema, CarIssue, \
 @api_blueprint.route('/admin/car/<int:car_id>/issue', methods=['POST'])
 @role_required(PersonType.ADMIN)
 def issue(car_id: int):
+    """
+    Endpoint to add/update/delete issue
+    Args:
+        car_id (int): car if for this issue (issue in one to one)
+
+    Returns: 403
+
+    """
+    car = Car.query.filter_by(id=car_id).first()
+    if car is None:
+        return abort(404, description='Car not found')
+
     schema = CarIssueSchema(exclude=['id'])
     try:
         issue = schema.loads(request.get_json())
     except ValidationError as ve:
-        return abort(400, description='Invalid issue data')
-    car = Car.query.filter_by(id=car_id).first()
+        return abort(400, description=ve.messages)
+
+    # issue text is removed -> delete issue from db
     if len(issue.issue) == 0:
         CarIssue.query.filter_by(id=car.issue.id).delete()
         db.session.commit()
         return jsonify('Issue deleted!'), 200
 
+    # add or update car issue
     db.session.add(issue)
     car.issue = issue
     db.session.commit()
@@ -32,23 +46,31 @@ def issue(car_id: int):
 @api_blueprint.route('/admin/car/<int:id>', methods=['GET', 'PUT', 'DELETE'])
 @role_required(PersonType.ADMIN)
 def car(id: int):
-    schema = CarSchema()
+    """
+    Get, update, or delete car
+    Args:
+        id (int): car id
+
+    Returns: Error or 200
+
+    """
     car = Car.query.get(id)
     if car is None:
         return abort(404, description='Car not found')
+    schema = CarSchema()
+
     if request.method == 'GET':
         # Dump to dict, then add issue text
         car_data = schema.dump(car)
         if car.issue is not None:
             car_data['issue'] = car.issue.issue
-
         return jsonify(json.dumps(car_data)), 200
 
     elif request.method == 'PUT':
         try:
             new_car = schema.loads(request.get_json())
         except ValidationError as ve:
-            return abort(400, description='Invalid car data')
+            return abort(400, description=ve.messages)
 
         car.reg_number = new_car.reg_number
         car.car_manufacturer = new_car.car_manufacturer
@@ -71,7 +93,11 @@ def car(id: int):
 @api_blueprint.route('/admin/car', methods=['GET'])
 @role_required(PersonType.ADMIN)
 def get_all_cars():
+    """
+    Endpoint to get all cars with extra data
+    Returns: car data
 
+    """
     cars = Car.query.all()
     car_data = [{
         'car': CarSchema().dump(car),
@@ -86,7 +112,11 @@ def get_all_cars():
 @api_blueprint.route('/admin/car', methods=['POST'])
 @role_required(PersonType.ADMIN)
 def add_car():
+    """
+    Endpoint to add car to db
+    Returns: car added to db, or 400
 
+    """
     schema = CarSchema(exclude=['id'])
     try:
         new_car = schema.loads(request.get_json())
@@ -95,5 +125,4 @@ def add_car():
 
     db.session.add(new_car)
     db.session.commit()
-
     return jsonify(schema.dumps(new_car)), 201
